+++ modules/module-postgres/package.json
@@ -13,7 +13,7 @@
     "build": "tsc -b",
     "build:tests": "tsc -b test/tsconfig.json",
     "clean": "rm -rf ./lib && tsc -b --clean",
-    "test": "vitest --no-threads"
+    "test": "vitest"
   },
   "exports": {
     ".": {
