+++ modules/module-postgres/src/utils/populate_test_data.ts
@@ -1,5 +1,4 @@
-import * as crypto from 'crypto';
-import { Worker, isMainThread, parentPort, workerData } from 'node:worker_threads';
+import { Worker } from 'node:worker_threads';
 
 import * as pgwire from '@powersync/service-jpgwire';
 
@@ -12,49 +11,10 @@ export interface PopulateDataOptions {
   size: number;
 }
 
-if (isMainThread || parentPort == null) {
-  // Not a worker - ignore
-} else {
-  try {
-    const options = workerData as PopulateDataOptions;
-
-    const result = await populateDataInner(options);
-    parentPort.postMessage(result);
-    process.exit(0);
-  } catch (e) {
-    // This is a bug, not a connection issue
-    console.error(e);
-    // Only closes the Worker thread
-    process.exit(2);
-  }
-}
-
-async function populateDataInner(options: PopulateDataOptions) {
-  // Dedicated connection so we can release the memory easily
-  const initialDb = await pgwire.connectPgWire(options.connection, { type: 'standard' });
-  const largeDescription = crypto.randomBytes(options.size / 2).toString('hex');
-  let operation_count = 0;
-  for (let i = 0; i < options.num_transactions; i++) {
-    const prefix = `test${i}K`;
-
-    await initialDb.query({
-      statement: `INSERT INTO test_data(id, description, other) SELECT $1 || i, $2, 'foo' FROM generate_series(1, $3) i`,
-      params: [
-        { type: 'varchar', value: prefix },
-        { type: 'varchar', value: largeDescription },
-        { type: 'int4', value: options.per_transaction }
-      ]
-    });
-    operation_count += options.per_transaction;
-  }
-  await initialDb.end();
-  return operation_count;
-}
-
 export async function populateData(options: PopulateDataOptions) {
   const WORKER_TIMEOUT = 30_000;
 
-  const worker = new Worker(new URL('./populate_test_data.js', import.meta.url), {
+  const worker = new Worker(new URL('./populate_test_data_worker.js', import.meta.url), {
     workerData: options
   });
   const timeout = setTimeout(() => {
