+++ packages/service-core/vitest.config.ts
@@ -4,6 +4,12 @@ import tsconfigPaths from 'vite-tsconfig-paths';
 export default defineConfig({
   plugins: [tsconfigPaths()],
   test: {
-    setupFiles: './test/src/setup.ts'
+    setupFiles: './test/src/setup.ts',
+    poolOptions: {
+      threads: {
+        singleThread: true
+      }
+    },
+    pool: 'threads'
   }
 });
