+++ test-client/package.json
@@ -20,7 +20,7 @@
     "yaml": "^2.5.0"
   },
   "devDependencies": {
-    "@types/node": "18.11.11",
+    "@types/node": "^22.5.5",
     "typescript": "^5.6.2"
   }
 }
